--- src/components/Forecast/Forecast.jsx
@@ -1,3 +1,0 @@
-
-import React from 'react';
-import App from '../../App';
